requirements.txt
pandas==1.2.4
numpy==1.19.5
scikit-learn==0.24.2
scipy==1.6.3
progressbar==2.5
lightgbm==3.2.1
xgboost==1.4.2
plotly==4.6.0
unidecode
fastapi==0.78.0
uvicorn==0.17.6
boto3==1.24.11
python-dotenv